Refactor first 1/3 of test_document_store.py
